api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
redis==6.1.0
python-dotenv==1.1.0